[project_member drop] 非推奨のメソッドを使わないようにうすｒ
