+++ annofabcli/project_member/subcommand_project_member.py
@@ -4,8 +4,8 @@ import annofabcli
 import annofabcli.common.cli
 import annofabcli.project_member.change_project_members
 import annofabcli.project_member.copy_project_members
-import annofabcli.project_member.delete_users
-import annofabcli.project_member.invite_users
+import annofabcli.project_member.drop_project_members
+import annofabcli.project_member.invite_project_members
 import annofabcli.project_member.list_users
 import annofabcli.project_member.put_project_members
 
@@ -17,8 +17,8 @@ def parse_args(parser: argparse.ArgumentParser):
     # サブコマンドの定義
     annofabcli.project_member.change_project_members.add_parser(subparsers)
     annofabcli.project_member.copy_project_members.add_parser(subparsers)
-    annofabcli.project_member.delete_users.add_parser(subparsers)
-    annofabcli.project_member.invite_users.add_parser(subparsers)
+    annofabcli.project_member.drop_project_members.add_parser(subparsers)
+    annofabcli.project_member.invite_project_members.add_parser(subparsers)
     annofabcli.project_member.list_users.add_parser(subparsers)
     annofabcli.project_member.put_project_members.add_parser(subparsers)
